@@ -2,18 +2,25 @@ package org.genericsystem.core;
 
 import java.util.List;
 import java.util.Set;
+import java.util.Spliterator;
+import java.util.Spliterators;
 
 /**
  * The Snaphot is an aware iterable of the graphe.
- * 
+ *
  * @author Nicolas Feybesse
- * 
+ *
  */
 public interface Snapshot<T> extends List<T>, Set<T> {
 
+	@Override
+	public default Spliterator<T> spliterator() {
+		return Spliterators.spliterator(this, Spliterator.ORDERED);
+	}
+
 	/**
 	 * Filter the Snapshot.
-	 * 
+	 *
 	 * @param filter
 	 *            The Filter.
 	 * @see Filter
@@ -23,7 +30,7 @@ public interface Snapshot<T> extends List<T>, Set<T> {
 
 	/**
 	 * Project the Snapshot.
-	 * 
+	 *
 	 * @param projector
 	 *            The Projecter.
 	 * @return The project Snapshot.
@@ -32,13 +39,14 @@ public interface Snapshot<T> extends List<T>, Set<T> {
 
 	/**
 	 * Filter.
-	 * 
+	 *
 	 * @author Nicolas Feybesse
 	 */
+	@FunctionalInterface
 	static interface Filter<T> {
 		/**
 		 * Returns true if element is selected.
-		 * 
+		 *
 		 * @param element
 		 *            The element.
 		 * @return True if element is selected.
@@ -48,13 +56,14 @@ public interface Snapshot<T> extends List<T>, Set<T> {
 
 	/**
 	 * Projector.
-	 * 
+	 *
 	 * @author Nicolas Feybesse
 	 */
+	@FunctionalInterface
 	static interface Projector<T, E> {
 		/**
 		 * Returns project element.
-		 * 
+		 *
 		 * @param element
 		 *            The initial element.
 		 * @return The project element.
@@ -63,4 +72,5 @@ public interface Snapshot<T> extends List<T>, Set<T> {
 	}
 
 	void log();
+
 }
