@@ -12,7 +12,6 @@ import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-
 import org.genericsystem.annotation.InstanceGenericClass;
 import org.genericsystem.annotation.NoInheritance;
 import org.genericsystem.annotation.SystemGeneric;
@@ -31,7 +30,6 @@ import org.genericsystem.constraints.SingularConstraintImpl;
 import org.genericsystem.constraints.SizeConstraintImpl;
 import org.genericsystem.constraints.UniqueValueConstraintImpl;
 import org.genericsystem.constraints.VirtualConstraintImpl;
-import org.genericsystem.core.Snapshot.Projector;
 import org.genericsystem.core.UnsafeGList.Components;
 import org.genericsystem.core.UnsafeGList.Supers;
 import org.genericsystem.core.UnsafeGList.UnsafeComponents;
@@ -70,7 +68,7 @@ import org.slf4j.LoggerFactory;
 /**
  * @author Nicolas Feybesse
  * @author Michael Ory
- * 
+ *
  */
 @SuppressWarnings("unchecked")
 public class GenericImpl implements Generic, Type, Link, Relation, Holder, Attribute {
@@ -263,13 +261,8 @@ public class GenericImpl implements Generic, Type, Link, Relation, Holder, Attri
 	}
 
 	@Override
-	public <T extends Serializable> Snapshot<T> getValues(final Holder attribute) {
-		return getHolders(attribute).project(new Projector<T, Holder>() {
-			@Override
-			public T project(Holder holder) {
-				return holder.<T> getValue();
-			}
-		});
+	public <T extends Serializable> AbstractSnapshot<T> getValues(final Holder attribute) {
+		return getHolders(attribute).project(holder -> holder.<T> getValue());
 	}
 
 	@Override
@@ -410,33 +403,23 @@ public class GenericImpl implements Generic, Type, Link, Relation, Holder, Attri
 	}
 
 	@Override
-	public <T extends Holder> Snapshot<T> getHolders(final Holder attribute, final Generic... targets) {
+	public <T extends Holder> AbstractSnapshot<T> getHolders(final Holder attribute, final Generic... targets) {
 		return getHolders(attribute, getBasePos(attribute), targets);
 	}
 
 	@Override
-	public <T extends Holder> Snapshot<T> getHolders(final Holder attribute, final int basePos, final Generic... targets) {
-		return new AbstractSnapshot<T>() {
-			@Override
-			public Iterator<T> iterator() {
-				return holdersIterator((Attribute) attribute, Statics.CONCRETE, basePos, targets);
-			}
-		};
+	public <T extends Holder> AbstractSnapshot<T> getHolders(final Holder attribute, final int basePos, final Generic... targets) {
+		return () -> holdersIterator((Attribute) attribute, Statics.CONCRETE, basePos, targets);
 	}
 
 	@Override
-	public <T extends Link> Snapshot<T> getLinks(final Relation relation, final Generic... targets) {
+	public <T extends Link> AbstractSnapshot<T> getLinks(final Relation relation, final Generic... targets) {
 		return getLinks(relation, getBasePos(relation), targets);
 	}
 
 	@Override
-	public <T extends Link> Snapshot<T> getLinks(final Relation relation, final int basePos, final Generic... targets) {
-		return new AbstractSnapshot<T>() {
-			@Override
-			public Iterator<T> iterator() {
-				return linksIterator(relation, basePos, targets);
-			}
-		};
+	public <T extends Link> AbstractSnapshot<T> getLinks(final Relation relation, final int basePos, final Generic... targets) {
+		return () -> linksIterator(relation, basePos, targets);
 	}
 
 	@Override
@@ -445,13 +428,8 @@ public class GenericImpl implements Generic, Type, Link, Relation, Holder, Attri
 	}
 
 	@Override
-	public <T extends Generic> Snapshot<T> getTargets(Relation relation, int basePos, final int targetPos) {
-		return getLinks(relation, basePos).project(new Projector<T, Link>() {
-			@Override
-			public T project(Link element) {
-				return element.getComponent(targetPos);
-			}
-		});
+	public <T extends Generic> AbstractSnapshot<T> getTargets(Relation relation, int basePos, final int targetPos) {
+		return getLinks(relation, basePos).project(element -> element.getComponent(targetPos));
 	}
 
 	public <T extends Holder> Iterator<T> holdersIterator(Holder attribute, Generic... targets) {
@@ -517,13 +495,8 @@ public class GenericImpl implements Generic, Type, Link, Relation, Holder, Attri
 	}
 
 	@Override
-	public <T extends Generic> Snapshot<T> getInheritings() {
-		return new AbstractSnapshot<T>() {
-			@Override
-			public Iterator<T> iterator() {
-				return inheritingsIterator();
-			}
-		};
+	public <T extends Generic> AbstractSnapshot<T> getInheritings() {
+		return () -> inheritingsIterator();
 	}
 
 	public <T extends Generic> Iterator<T> compositesIterator() {
@@ -540,13 +513,8 @@ public class GenericImpl implements Generic, Type, Link, Relation, Holder, Attri
 	}
 
 	@Override
-	public <T extends Generic> Snapshot<T> getComposites() {
-		return new AbstractSnapshot<T>() {
-			@Override
-			public Iterator<T> iterator() {
-				return compositesIterator();
-			}
-		};
+	public <T extends Generic> AbstractSnapshot<T> getComposites() {
+		return () -> compositesIterator();
 	}
 
 	public <T extends Generic> Iterator<T> attributesIterator() {
@@ -554,23 +522,13 @@ public class GenericImpl implements Generic, Type, Link, Relation, Holder, Attri
 	}
 
 	@Override
-	public <T extends Attribute> Snapshot<T> getAttributes() {
-		return new AbstractSnapshot<T>() {
-			@Override
-			public Iterator<T> iterator() {
-				return attributesIterator();
-			}
-		};
+	public <T extends Attribute> AbstractSnapshot<T> getAttributes() {
+		return () -> attributesIterator();
 	}
 
 	@Override
-	public <T extends Attribute> Snapshot<T> getAttributes(final Attribute attribute) {
-		return new AbstractSnapshot<T>() {
-			@Override
-			public Iterator<T> iterator() {
-				return holdersIterator(Statics.STRUCTURAL, attribute, Statics.MULTIDIRECTIONAL);
-			}
-		};
+	public <T extends Attribute> AbstractSnapshot<T> getAttributes(final Attribute attribute) {
+		return () -> holdersIterator(Statics.STRUCTURAL, attribute, Statics.MULTIDIRECTIONAL);
 	}
 
 	@Override
@@ -962,12 +920,7 @@ public class GenericImpl implements Generic, Type, Link, Relation, Holder, Attri
 		final Iterable<Generic>[] projections = new Iterable[getComponents().size()];
 		for (int i = 0; i < projections.length; i++) {
 			final int column = i;
-			projections[i] = new Iterable<Generic>() {
-				@Override
-				public Iterator<Generic> iterator() {
-					return pos != column && getComponents().get(column).isStructural() ? ((GenericImpl) getComponents().get(column)).allInstancesIterator() : new SingletonIterator<Generic>(getComponents().get(column));
-				}
-			};
+			projections[i] = () -> pos != column && getComponents().get(column).isStructural() ? ((GenericImpl) getComponents().get(column)).allInstancesIterator() : new SingletonIterator<Generic>(getComponents().get(column));
 		}
 		return projections;
 	}
@@ -1382,13 +1335,8 @@ public class GenericImpl implements Generic, Type, Link, Relation, Holder, Attri
 	}
 
 	@Override
-	public <T extends Generic> Snapshot<T> getSubTypes() {
-		return new AbstractSnapshot<T>() {
-			@Override
-			public Iterator<T> iterator() {
-				return subTypesIterator();
-			}
-		};
+	public <T extends Generic> AbstractSnapshot<T> getSubTypes() {
+		return () -> subTypesIterator();
 	}
 
 	private <T extends Generic> Iterator<T> subTypesIterator() {
@@ -1396,13 +1344,8 @@ public class GenericImpl implements Generic, Type, Link, Relation, Holder, Attri
 	}
 
 	@Override
-	public <T extends Generic> Snapshot<T> getAllSubTypes() {
-		return new AbstractSnapshot<T>() {
-			@Override
-			public Iterator<T> iterator() {
-				return allSubTypesIteratorWithoutRoot();
-			}
-		};
+	public <T extends Generic> AbstractSnapshot<T> getAllSubTypes() {
+		return () -> allSubTypesIteratorWithoutRoot();
 	}
 
 	// TODO super KK what is this method, what dost it do : no components ? no supers ? ???
@@ -1412,38 +1355,20 @@ public class GenericImpl implements Generic, Type, Link, Relation, Holder, Attri
 	}
 
 	@Override
-	public <T extends Generic> Snapshot<T> getAllSubTypes(final String name) {
-		return new AbstractSnapshot<T>() {
-
-			@Override
-			public Iterator<T> iterator() {
-				return Statics.valueFilter(GenericImpl.this.<T> allSubTypesIteratorWithoutRoot(), name);
-			}
-		};
+	public <T extends Generic> AbstractSnapshot<T> getAllSubTypes(final String name) {
+		return () -> Statics.valueFilter(GenericImpl.this.<T> allSubTypesIteratorWithoutRoot(), name);
 	}
 
 	private <T extends Generic> Iterator<T> allSubTypesIteratorWithoutRoot() {
 		return Statics.levelFilter(this.<T> allInheritingsIteratorWithoutRoot(), Statics.STRUCTURAL);
 	}
 
-	public <T extends Generic> Snapshot<T> getAllInheritings() {
-		return new AbstractSnapshot<T>() {
-
-			@Override
-			public Iterator<T> iterator() {
-				return allInheritingsIterator();
-			}
-		};
+	public <T extends Generic> AbstractSnapshot<T> getAllInheritings() {
+		return () -> allInheritingsIterator();
 	}
 
-	public <T extends Generic> Snapshot<T> getAllInheritingsWithoutRoot() {
-		return new AbstractSnapshot<T>() {
-
-			@Override
-			public Iterator<T> iterator() {
-				return allInheritingsIteratorWithoutRoot();
-			}
-		};
+	public <T extends Generic> AbstractSnapshot<T> getAllInheritingsWithoutRoot() {
+		return () -> allInheritingsIteratorWithoutRoot();
 	}
 
 	protected <T extends Generic> Iterator<T> allInheritingsIterator() {
@@ -1925,13 +1850,8 @@ public class GenericImpl implements Generic, Type, Link, Relation, Holder, Attri
 		}
 	}
 
-	public Snapshot<Integer> getPositions(final Holder attribute) {
-		return new AbstractSnapshot<Integer>() {
-			@Override
-			public Iterator<Integer> iterator() {
-				return positionsIterator(attribute);
-			}
-		};
+	public AbstractSnapshot<Integer> getPositions(final Holder attribute) {
+		return () -> positionsIterator(attribute);
 	}
 
 	Iterator<Integer> positionsIterator(final Holder attribute) {
@@ -1945,13 +1865,8 @@ public class GenericImpl implements Generic, Type, Link, Relation, Holder, Attri
 	}
 
 	@Override
-	public Snapshot<Generic> getOtherTargets(final Holder holder) {
-		return new AbstractSnapshot<Generic>() {
-			@Override
-			public Iterator<Generic> iterator() {
-				return otherTargetsIterator(holder);
-			}
-		};
+	public AbstractSnapshot<Generic> getOtherTargets(final Holder holder) {
+		return () -> otherTargetsIterator(holder);
 	}
 
 	public Iterator<Generic> otherTargetsIterator(final Holder holder) {
